+++ inst/templates/unit_tests/tests/testthat/test-main.R
@@ -1,9 +1,9 @@
 box::use(
   shiny[testServer],
-  testthat[...],
+  testthat[expect_true, test_that],
 )
 box::use(
-  app/main[...],
+  app/main[server, ui],
 )
 
 test_that("main server works", {